Review Refactor
- Refactor Review Flow to use SDK and cli args.
- Update Tenet bundles to new syntax
- Add variable flag to review and rewrite
