+++ tenets/codelingo/go/println-format-strings/codelingo.yaml
@@ -11,7 +11,7 @@ tenets:
     query: |
       import codelingo/ast/go
 
-      @review.comment
+      @review comment
       go.call_expr(depth = any):
         go.selector_expr:
           go.ident:
