+++ tenets/codelingo/go/todo/codelingo.yaml
@@ -14,7 +14,7 @@ tenets:
         codelingo/ast/go
       )
 
-      @ review.comment
+      @review comment
       go.comment(depth = any):
         text as todoText
         regex(/(?i)(^|\s)TODO.*/, todoText)
