+++ tenets/codelingo/jenkinsx/parallel-in-tests/codelingo.yaml
@@ -22,7 +22,7 @@ tenets:
       go.file(depth = any):
         filename as fname
         excludeIntegration(fname)
-        @review.comment
+        @review comment
         go.func_decl(depth = any):
           go.ident:
             name as funcName
