+++ tenets/codelingo/jenkinsx/test-package-name/codelingo.yaml
@@ -20,7 +20,7 @@ tenets:
       import codelingo/ast/go
 
       go.file(depth = any):
-        @review.comment
+        @review comment
         go.ident:
           name as packageName
           missingTestSuffix(packageName)
