+++ tenets/codelingo/php/empty-catch-block/codelingo.yaml
@@ -10,6 +10,6 @@ tenets:
     query: |
       import codelingo/ast/php
 
-      @ review.comment
+      @review comment
       php.stmt_catch(depth = any):
         child_count == "1"
