+++ tenets/codelingo/php/if-assignment/codelingo.yaml
@@ -10,6 +10,6 @@ tenets:
     query: |
       import codelingo/ast/php
 
-      @ review.comment
+      @review comment
       php.stmt_if(depth = any):
         php.expr_assign
