+++ tenets/codelingo/php/sql-concats/codelingo.yaml
@@ -11,12 +11,12 @@ tenets:
       php.file(depth = any):
         any_of:
           php.expr_assignop_concat(depth = any):
-            @ review.comment
+            @review comment
             php.expr_variable(depth = any):
               name as sqlName1
               regex(/(?i)SQL/, sqlName1)
           php.expr_assign(depth = any):
-            @ review.comment
+            @review comment
             php.expr_variable(depth = any):
               name as sqlName2
               regex(/(?i)SQL/, sqlName2)
