+++ tenets/west/default/shadowing/codelingo.yaml
@@ -13,7 +13,7 @@ tenets:
         cpp.parm_decl:
           name as paramName
         cpp.compound_stmt:
-          @ review.comment
+          @review comment
           cpp.binary_operator(depth = any):
             cpp.decl_ref_expr:
               name as paramName
@@ -32,7 +32,7 @@ tenets:
         cpp.parm_decl:
           name as varName
       cpp.compound_stmt(depth = any):
-        @ review.comment
+        @review comment
         cpp.binary_operator(depth = any):
           cpp.decl_ref_expr:
             name as varName
